ai_chat_bot/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
supabase
python-dotenv
pyjwt
cryptography
requests
